날씨 서비스 형식 일부 수정

+++ app/src/main/java/com/teamhub/notiget/model/weather/openweather/DailyModel.java
@@ -9,7 +9,7 @@ public class DailyModel {
     public long dt;
     public FeelsLikeModel feels_like;
     public int humidity;
-    public int pop;
+    public double pop;
     public int pressure;
     public long sunrise;
     public long sunset;
